+++ 릴규-이정규/MovieAPI/src/App.jsx
@@ -2,11 +2,11 @@ import React from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Navbar from "./components/Navbar";
 import Footer from "./components/Footer";
-import MainPage from "./components/MainPage";
+import MainPage from "./pages/MainPage";
 import SignUpPage from "./components/SignUpPage";
-import PopularPage from "./components/PopularPage";
-import NowPlayingPage from "./components/NowPlayingPage";
-import TopRatedPage from "./components/TopRatedPage";
+import PopularPage from "./pages/PopularPage";
+import NowPlayingPage from "./pages/NowPlayingPage";
+import TopRatedPage from "./pages/TopRatedPage";
 import UpcomingPage from "./components/UpcomingPage";
 import styled from "styled-components";
